[Add] フラッシュメッセージ

+++ app/controllers/books_controller.rb
@@ -31,7 +31,7 @@ class BooksController < ApplicationController
     if
       book.update(book_params)
       flash[:notice] = "Book was successfully updated."
-    redirect_to book_path(book.id)
+      redirect_to book_path(book.id)
     else
       render :new
     end
@@ -39,8 +39,13 @@ class BooksController < ApplicationController
   
   def destroy
     book = Book.find(params[:id])
-    book.destroy
-    redirect_to '/books'
+    if 
+      book.destroy
+      flash[:notice] = "Book was successfully destroyed."
+      redirect_to '/books'
+    else
+      render :new
+    end
   end
 
   private
